some small refactorings and more doc
